Add files via upload
This was a project done during my Summer 2022 internship using a Kaggle data set to understand the problems of customer retention. It was done to further investigate the different predictors that are most influential in customers' decisions to churn (or not). The findings of this project were used to make informed decisions in choosing pertinent and relevant predictors for executing the actual project with our client. Due to signing an NDA with the company, I am unable to disclose and post the full project we worked on during my time as an intern with the relational database provided by the said company.